Move EmitTestLogMsg/EmitTestExplain to the runner package
This avoids having the logs package depend on the input/postgres package,
which would have caused an import cycle in an upcoming commit.
