+++ runner/emit_test_lines.go
@@ -1,4 +1,4 @@
-package logs
+package runner
 
 import (
 	"fmt"
